requirements.txt
fastapi
uvicorn
azure-storage-blob
python-dotenv